collector/requirements_collector.txt
kafka-python==2.0.2
pymongo==4.6.1
requests==2.31.0
lxml==4.9.4
feedparser==6.0.11
python-dotenv==1.0.0
loguru==0.7.2
schedule==1.2.1